requirements.txt
Flask==3.1.0
flask_sqlalchemy==3.1.1
groq==0.15.0
python-dotenv==1.0.1
